put log levels into parse_args.go; add godocs
