+++ parse_args.go
@@ -10,6 +10,12 @@ import (
 	"strings"
 )
 
+// Arg represents an argument passed to the program at initial execution.
+// It has a name and a value where an Arg can only have one value
+//
+// An Arg can be a command, a switch, or a flag.
+//	Commands generally have no dashes
+//	Switches and flags generally have a dash (-) prepended to their names
 type Arg struct {
 	name  string
 	value string
@@ -89,6 +95,7 @@ func TokenizeArgs(args []string) ([]Arg, error) {
 	return tokenizedArgs, nil
 }
 
+// Args is a list of arguments needed by gorn to properly rename media files
 type Args struct {
 	root    []string
 	series  []string
@@ -96,12 +103,76 @@ type Args struct {
 	options Flags
 	log 	LogFlag
 }
+// Flags are options for modifying the behavior of renaming files
 type Flags struct {
 	keepEpNums    Option[bool]
 	startingEpNum Option[int]
 	hasSeason0    Option[bool]
 	namingScheme  Option[string]
 }
+
+// LogFlag is a flag specifically for logging.
+// It handles which logs to print based on level
+type LogFlag struct {
+	level LogLevel
+}
+// For text color on log headers
+const (
+	// for informational logs
+	INFO = "[INFO] " // no color
+
+	// can safely skip error, doesn't interrupt process
+	WARN = "\033[93m[WARN]\033[0m " // yellow
+
+	// cannot safely skip error, must interrupt process
+	FATAL = "\033[91m[FATAL]\033[0m " // red
+
+	// for timing purposes
+	TIME = "\033[94m[TIME]\033[0m " // blue
+)
+type LogLevel int8 // can only be 1-4
+const (
+	FATAL_LEVEL LogLevel = iota + 1
+	WARN_LEVEL
+	INFO_LEVEL
+	TIME_LEVEL
+)
+func (l *LogFlag) Level() (string, error) {
+	switch l.level {
+	case FATAL_LEVEL:
+		return fmt.Sprintln(FATAL), nil
+	case WARN_LEVEL:
+		return fmt.Sprintln(FATAL, WARN), nil
+	case INFO_LEVEL:
+		return fmt.Sprintln(FATAL, WARN, INFO), nil
+	case TIME_LEVEL:
+		return fmt.Sprintln(FATAL, WARN, INFO, TIME), nil
+	default:
+		return "", fmt.Errorf("invalid log level: %d", l.level)
+	}
+}
+// ToLogLevel converts a string to a LogLevel if the string passed is valid; otherwise it returns an error.
+// 
+// Valid log levels:
+//	none, all, fatal, warn, info, time
+func ToLogLevel(s string) (LogLevel, error) {
+	switch s {
+	case "fatal":
+		return FATAL_LEVEL, nil
+	case "warn":
+		return WARN_LEVEL, nil
+	case "info", "", "all":
+		return INFO_LEVEL, nil
+	case "time":
+		return TIME_LEVEL, nil
+	case "none":
+		return 0, nil
+	}
+	return -1, fmt.Errorf("invalid value '%s' for --logs. Must be 'all', 'none', or a valid log header", s)
+}
+// newArgs returns a new Args struct with default values for the Flags
+//
+// Commands though are empty by default and need to be populated
 func newArgs() Args {
 	return Args{
 		root:   make([]string, 0),
@@ -354,7 +425,10 @@ func ParseArgs(args []Arg) (Args, error) {
 
 	return parsedArgs, nil
 }
-
+// ValidateNamingScheme checks if a naming scheme is valid by:
+//  - tokenizing APIs
+//  - checking if each API is valid
+//  - validating each API's value if user provided any
 func ValidateNamingScheme(s string) error {
 	if s[0] != '"' || s[len(s)-1] != '"' {
 		return fmt.Errorf("naming scheme must be enclosed in double quotes: %s", s)
@@ -453,7 +527,7 @@ func ValidateNamingScheme(s string) error {
 
 	return nil
 }
-
+// Splits the naming scheme into a list of APIs
 func TokenizeNamingScheme(s string) ([]string, error) {
 	isToken := false
 	builder := strings.Builder{}
@@ -488,7 +562,11 @@ func TokenizeNamingScheme(s string) ([]string, error) {
 
 	return namingScheme, nil
 }
-
+// ValidateRoots checks if:
+//  - there is at least one root/source directory
+//  - each root/source directory exists
+//  - each root/source directory is not a subdirectory of another root/source directory
+//  - each root/source directory is not a duplicate of another root/source directory
 func ValidateRoots(root []string, series []string, movies []string) error {
 	// must at least have one of any
 	if len(root) == 0 && len(series) == 0 && len(movies) == 0 {
